DT-88
DT-88: As a user, I'd like to have access to a quick link for the Youtube and Facebook pages of the church.

Description: Append Facebook and YouTube links to template

Acceptance Criteria: Use respective images/logo for Facebook and YouTube and link the church's pages in the. Add alt attribute to let screen readers and braille monitors know what they are

Personal Notes:
- All HTML files now use new footer w/ YouTube, FaceBook links
- New file: footer.css
- Moved some code from header.css to footer.css
- New images added to imagesWeb folder

Note: The YouTube and Facebook links are placeholders